Finally got the flow owrking for mac

+++ chicken.h
@@ -31,8 +31,7 @@
 #ifndef ___CHICKEN
 #define ___CHICKEN
 
-void setpixelcb(void (*newcb)(int, int, int, int, int, int));
-// void setpixels(char *newpixels);
+void setpixelcb(void (*newcb)(int, int, uint8_t, uint8_t, uint8_t, uint8_t));
 
 #define C_MAJOR_VERSION   5
 #define C_MINOR_VERSION   2
